+++ packages/base-editor/src/managers/EditorContext.js
@@ -3,6 +3,7 @@ import { ConfigManager } from "./ConfigManager";
 import { PluginManager } from "./PluginManager";
 import { UIManager } from "./UIManager";
 
+const CONTEXT_ID = "EditorContext";
 export class EditorContext {
   constructor($rootEditor, $options = {}) {
     this.$rootEditor = $rootEditor;
@@ -32,6 +33,10 @@ export class EditorContext {
       commands: CommandManager,
       plugins: PluginManager,
       uis: UIManager,
+      //TODO: i18n: I18nManager,            // 다국어 매니저
+      //TODO: history: HistoryManager,      // history manager
+      //TODO: inspector: InspectorManager,  // inspector 관리용 매니저
+      //TODO: state: StateManager,  // config 와 관계없는 에디터 공용 캐쉬 저장소
       ...managers,
     };
 
@@ -53,13 +58,13 @@ export class EditorContext {
 
   initializeConfigs(configs = []) {
     configs.forEach((config) => {
-      this.configs.registerPlugin(config);
+      this.configs.registerConfig(config);
     });
   }
 
   initializeCommands(commands = []) {
     commands.forEach((command) => {
-      this.commands.registerPlugin(command);
+      this.commands.registerCommand(command);
     });
   }
 
@@ -78,6 +83,7 @@ export class EditorContext {
   }
 
   emit(message, ...args) {
+    this.$store.source = CONTEXT_ID;
     this.$store.emit(message, ...args);
   }
 
@@ -89,7 +95,23 @@ export class EditorContext {
     this.uis.registerUI(ui);
   }
 
+  registerGroupUI(group, ui) {
+    this.uis.registerGroupUI(group, ui);
+  }
+
+  registerConfig(config) {
+    this.configs.registerConfig(config);
+  }
+
   getUI(name) {
     return this.uis.getUI(name);
+  }
+
+  getGroupUI(group) {
+    return this.uis.getGroupUI(group);
+  }
+
+  getConfig(key) {
+    return this.configs.get(key);
   }
 }